Development

+++ backend/test_app.py
@@ -7,7 +7,6 @@ from flask_sqlalchemy import SQLAlchemy
 from app import create_app
 from models import setup_db, Artist, Client, Appointment
 
-
 class TattooShopTestCase(unittest.TestCase):
     # This class represents the tattoo test case
 
@@ -19,6 +18,9 @@ class TattooShopTestCase(unittest.TestCase):
         self.database_path = "postgresql://postgres@localhost:5432/{}".format(
                                 self.database_name
                                 )
+        self.manager_jwt = 'Bearer eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCIsImtpZCI6Inh0Z1Q3dl95ZzB0VURmV2VnYnBhXyJ9.eyJpc3MiOiJodHRwczovL2ZzbmQtOC51cy5hdXRoMC5jb20vIiwic3ViIjoiZ29vZ2xlLW9hdXRoMnwxMDAzNzMzMTE4MzgwMTczNjk1MzkiLCJhdWQiOlsiaHR0cHM6Ly90YXR0b28tYXBpIiwiaHR0cHM6Ly9mc25kLTgudXMuYXV0aDAuY29tL3VzZXJpbmZvIl0sImlhdCI6MTYwOTAyMDk1OCwiZXhwIjoxNjA5MDI4MTU4LCJhenAiOiJQcE0yT0VwbFZKbEV5RTV4eDN2TFNiN1JtTW1RZEYxQyIsInNjb3BlIjoib3BlbmlkIHByb2ZpbGUgZW1haWwiLCJwZXJtaXNzaW9ucyI6WyJjcmVhdGU6YXBwb2ludG1lbnQiLCJjcmVhdGU6YXJ0aXN0IiwiY3JlYXRlOmNsaWVudCIsImRlbGV0ZTphcHBvaW50bWVudCIsImRlbGV0ZTphcnRpc3QiLCJkZWxldGU6Y2xpZW50IiwiZ2V0OmFsbCIsImdldDphcHBvaW50bWVudCIsInVwZGF0ZTphcHBvaW50bWVudCIsInVwZGF0ZTphcnRpc3QiLCJ1cGRhdGU6Y2xpZW50Il19.ojzI5cNByKJ74zf9ntcxO0FkgDR8NapVe6Z8afQ74CbPyzTtM1ACA2CT68fr6OZgt94r_KxXnZAN8gxmOJoyC6XPeI9CqyjNviwH_bm5KkbrQY7bmuubEyND_L170_nO_7KmhyhokQfAVIlH0XLtEq6zLscnxp9GCN6Pad4SuDazf3S5iwgSkSmArFbvAMiJGokM00bu1R6_8cKA4w_Ilx-yrojWt0XrmxOHn8HuAVHkJQGf30dtQat3gVUh8BpDakMsVSO10mkCkhMb3ZCpoDNa8xIaSxbs1v2c_1Dggg1OnDYiqhEryY6KHZLadj431O-bGQZt-Zri8PBb1ZvAwg'
+        self.client_jwt = 'Bearer eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCIsImtpZCI6Inh0Z1Q3dl95ZzB0VURmV2VnYnBhXyJ9.eyJpc3MiOiJodHRwczovL2ZzbmQtOC51cy5hdXRoMC5jb20vIiwic3ViIjoiYXV0aDB8NWZlNTI3YWE0NTIyNjgwMDc1NWEzMjNhIiwiYXVkIjoiaHR0cHM6Ly90YXR0b28tYXBpIiwiaWF0IjoxNjA5MDIwODY0LCJleHAiOjE2MDkwMjgwNjQsImF6cCI6IlBwTTJPRXBsVkpsRXlFNXh4M3ZMU2I3Um1NbVFkRjFDIiwic2NvcGUiOiIiLCJwZXJtaXNzaW9ucyI6WyJnZXQ6YXBwb2ludG1lbnQiXX0.GXvLKAqttOGbk7QAxIJ6cZKpFbTgh86cPIPqI5I0GTY8-W0pQmGIA075dvbylL86QqJ3DaiBJaDIN2Rfc7Y0QmUQiGgBfOjrAj7SSaZtEqmlC-WsQaNmIKZwO2VJJQXw5HyWD6ePu7OT9p36SPqsxWCiB78ywgGv7R0r92ZC1G-_LwqwZqmuqtGp5NOr8IOS9AvHANhzgOF8NfbYChgSVqNAQIfoyVaztuYqHBhlyGQp4XWlSR-Fa3eW87EjtMJbJeDeOcZKn-kbVqLhO5CoFthFdxJNauZBIFDBooFaFRjx1JiRYuLJ5LQ-cvaUi--OYfO9hGphyD88SV5qDKvMPQ'
+        self.artist_jwt = 'Bearer eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCIsImtpZCI6Inh0Z1Q3dl95ZzB0VURmV2VnYnBhXyJ9.eyJpc3MiOiJodHRwczovL2ZzbmQtOC51cy5hdXRoMC5jb20vIiwic3ViIjoiYXV0aDB8NWZlNTJjMzBiYmM0ZjkwMDZmMWZhYjY3IiwiYXVkIjoiaHR0cHM6Ly90YXR0b28tYXBpIiwiaWF0IjoxNjA5MDIxMDMxLCJleHAiOjE2MDkwMjgyMzEsImF6cCI6IlBwTTJPRXBsVkpsRXlFNXh4M3ZMU2I3Um1NbVFkRjFDIiwic2NvcGUiOiIiLCJwZXJtaXNzaW9ucyI6WyJjcmVhdGU6YXBwb2ludG1lbnQiLCJjcmVhdGU6Y2xpZW50IiwiZGVsZXRlOmFwcG9pbnRtZW50IiwiZ2V0OmFsbCIsImdldDphcHBvaW50bWVudCIsInVwZGF0ZTphcHBvaW50bWVudCIsInVwZGF0ZTpjbGllbnQiXX0.Zjd58lCub0-PIlMvlxl9aWo5DvtcAeZg7cG8QcraKAJJ8ID_LxULaPedasrlYMk8w6Z02sH3aMVqp3tK6sZJs0CGk-YBzdcKnJuTwx_76-32-iQQow4-9otVRPn-zZ8F-Cm0yds4Mq2DnpK3YExtMpq2KxoHPWuSVUa4oZWqRo6lYAP4vtk-dtzF-iUp8IdvlI-6Y5lavi1kIkLV1MByPLxdbtnm8lXgBCxC-yKQAzekzeLN7IzGe89EXSp_4d-XzFhAozsE3-2T3QWCcP2E9hoNsBETrpDonUgLBjgMJKbK5JOJsqdAjKuHSuOuPtw3JrjflpyIXn2Kyh13lXPhmQ'
         setup_db(self.app, self.database_path)
 
         with self.app.app_context():
@@ -57,7 +59,7 @@ class TattooShopTestCase(unittest.TestCase):
     def test_get_all_clients(self):
         # Test GET all clients endpoint returns:
         # all clients and 200 OK status
-        res = self.client().get('/api/clients')
+        res = self.client().get('/api/clients', headers={"Authorization": self.manager_jwt})
         data = json.loads(res.data)
 
         self.assertEqual(res.status_code, 200)
@@ -68,7 +70,7 @@ class TattooShopTestCase(unittest.TestCase):
     def test_get_client_by_id(self):
         # Test GET client according to artist_id returns
         # client id and 200 OK status
-        res = self.client().get('/api/clients/2')
+        res = self.client().get('/api/clients/2', headers={"Authorization": self.manager_jwt})
         data = json.loads(res.data)
 
         self.assertEqual(res.status_code, 200)
@@ -79,7 +81,7 @@ class TattooShopTestCase(unittest.TestCase):
         # Test GET appointment according to appt id returns:
         # appt id and 200 OK
 
-        res = self.client().get('/api/appointments/2')
+        res = self.client().get('/api/appointments/2', headers={"Authorization": self.manager_jwt})
         data = json.loads(res.data)
 
         self.assertEqual(res.status_code, 200)
@@ -101,7 +103,8 @@ class TattooShopTestCase(unittest.TestCase):
                    'instagram_link': '',
                    'email': '',
                    }
-        res = self.client().post('/api/artists', json=payload)
+            
+        res = self.client().post('/api/artists', headers={"Authorization": self.manager_jwt}, json=payload)
         data = json.loads(res.data)
 
         self.assertEqual(res.status_code, 200)
@@ -119,7 +122,7 @@ class TattooShopTestCase(unittest.TestCase):
                    'email': '',
                    'address': '',
                    }
-        res = self.client().post('/api/clients', json=payload)
+        res = self.client().post('/api/clients', json=payload, headers={"Authorization": self.manager_jwt})
         data = json.loads(res.data)
 
         self.assertEqual(res.status_code, 200)
@@ -136,7 +139,7 @@ class TattooShopTestCase(unittest.TestCase):
                    'artist': 1,
                    'appointment_date': datetime(2021, 3, 6, 12, 30)
                    }
-        res = self.client().post('/api/appointments', json=payload)
+        res = self.client().post('/api/appointments', json=payload, headers={"Authorization": self.manager_jwt})
         data = json.loads(res.data)
 
         last_appt = Appointment.query.all()[-1].id
@@ -159,7 +162,7 @@ class TattooShopTestCase(unittest.TestCase):
                     'email': 'creative_genius@aol.com'
                    }
 
-        res = self.client().patch('/api/artists/2', json=payload)
+        res = self.client().patch('/api/artists/2', json=payload, headers={"Authorization": self.manager_jwt})
         data = json.loads(res.data)
 
         artist = Artist.query.get(2)
@@ -178,7 +181,7 @@ class TattooShopTestCase(unittest.TestCase):
                     'phone': '770-231-4234',
                     'email': 'simplord12@gmail.com'
                     }
-        res = self.client().patch('/api/clients/2', json=payload)
+        res = self.client().patch('/api/clients/2', json=payload, headers={"Authorization": self.manager_jwt})
         data = json.loads(res.data)
 
         self.assertEqual(res.status_code, 200)
@@ -193,7 +196,7 @@ class TattooShopTestCase(unittest.TestCase):
         payload = {
                     "appointment_date": datetime(2021, 8, 24, 2, 30)
                     }
-        res = self.client().patch('/api/appointments/1', json=payload)
+        res = self.client().patch('/api/appointments/1', json=payload, headers={"Authorization": self.manager_jwt})
         data = json.loads(res.data)
 
         self.assertEqual(res.status_code, 200)
@@ -210,7 +213,7 @@ class TattooShopTestCase(unittest.TestCase):
         # Test sending a delete method for an existing artist returns:
         # id of the deleted artist and 200 OK
         artist_id = Artist.query.order_by(Artist.id).all()[-1].id
-        res = self.client().delete(f"/api/artists/{artist_id}")
+        res = self.client().delete(f"/api/artists/{artist_id}", headers={"Authorization": self.manager_jwt})
         data = json.loads(res.data)
 
         self.assertEqual(res.status_code, 200)
@@ -222,7 +225,7 @@ class TattooShopTestCase(unittest.TestCase):
         # Test sending a delete method for an existing client returns:
         # id of the deleted client and 200 OK
         client_id = Client.query.order_by(Client.id).all()[-1].id
-        res = self.client().delete(f"/api/clients/{client_id}")
+        res = self.client().delete(f"/api/clients/{client_id}", headers={"Authorization": self.manager_jwt})
         data = json.loads(res.data)
 
         self.assertEqual(res.status_code, 200)
@@ -234,7 +237,7 @@ class TattooShopTestCase(unittest.TestCase):
         # Test sending a delete method for an existing appointment returns:
         # id of the deleted appointment and 200 OK
         appt_id = Appointment.query.order_by(Appointment.id).all()[-1].id
-        res = self.client().delete(f"/api/appointments/{appt_id}")
+        res = self.client().delete(f"/api/appointments/{appt_id}", headers={"Authorization": self.manager_jwt})
         data = json.loads(res.data)
 
         self.assertEqual(res.status_code, 200)
@@ -268,7 +271,7 @@ class TattooShopTestCase(unittest.TestCase):
     # Test request for clients page that does not exist returns 404
     def test_get_all_clients_error(self):
         
-        res = self.client().get('/api/clients?page=1000')
+        res = self.client().get('/api/clients?page=1000', headers={"Authorization": self.manager_jwt})
         data = json.loads(res.data)
 
         self.assertEqual(res.status_code, 404)
@@ -278,7 +281,7 @@ class TattooShopTestCase(unittest.TestCase):
     # Test request for client that does not exist returns 404
     def test_get_client_by_id_error(self):
         
-        res = self.client().get('/api/clients/10000')
+        res = self.client().get('/api/clients/10000', headers={"Authorization": self.manager_jwt})
         data = json.loads(res.data)
 
         self.assertEqual(res.status_code, 404)
@@ -288,7 +291,7 @@ class TattooShopTestCase(unittest.TestCase):
     # Test request for appointment that does not exist returns 404
     def test_get_appointment_by_id_error(self):
         
-        res = self.client().get('/api/appointments/10000')
+        res = self.client().get('/api/appointments/10000', headers={"Authorization": self.manager_jwt})
         data = json.loads(res.data)
 
         self.assertEqual(res.status_code, 404)
@@ -308,7 +311,7 @@ class TattooShopTestCase(unittest.TestCase):
                    'instagram_link': '',
                    'email': '',
                    }
-        res = self.client().post('/api/artists', json=payload)
+        res = self.client().post('/api/artists', json=payload, headers={"Authorization": self.manager_jwt})
         data = json.loads(res.data)
 
         self.assertEqual(res.status_code, 422)
@@ -322,7 +325,7 @@ class TattooShopTestCase(unittest.TestCase):
                    'email': '',
                    'address': '',
                    }
-        res = self.client().post('/api/clients', json=payload)
+        res = self.client().post('/api/clients', json=payload, headers={"Authorization": self.manager_jwt})
         data = json.loads(res.data)
 
         self.assertEqual(res.status_code, 422)
@@ -335,7 +338,7 @@ class TattooShopTestCase(unittest.TestCase):
                    'client': 1,
                    'artist': 1,
                    }
-        res = self.client().post('/api/appointments', json=payload)
+        res = self.client().post('/api/appointments', json=payload, headers={"Authorization": self.manager_jwt})
         data = json.loads(res.data)
 
         self.assertEqual(res.status_code, 422)
@@ -356,7 +359,7 @@ class TattooShopTestCase(unittest.TestCase):
                    'image_link': '',
                    'instagram_link': '',
                     }
-        res = self.client().patch('/api/artists/1000')
+        res = self.client().patch('/api/artists/1000', headers={"Authorization": self.manager_jwt})
         data = json.loads(res.data)
 
         self.assertEqual(res.status_code, 404)
@@ -371,7 +374,7 @@ class TattooShopTestCase(unittest.TestCase):
                    'email': '',
                    'address': '',
                    }
-        res = self.client().patch('/api/clients/1000', json=payload)
+        res = self.client().patch('/api/clients/1000', json=payload, headers={"Authorization": self.manager_jwt})
         data = json.loads(res.data)
 
         self.assertEqual(res.status_code, 404)
@@ -384,7 +387,7 @@ class TattooShopTestCase(unittest.TestCase):
                     'appointment_date': "2020, 12 31, 12:30:54"
                     }
 
-        res = self.client().patch('/api/appointments/2', json=payload)
+        res = self.client().patch('/api/appointments/2', json=payload, headers={"Authorization": self.manager_jwt})
         data = json.loads(res.data)
 
         self.assertEqual(res.status_code, 422)
@@ -398,7 +401,7 @@ class TattooShopTestCase(unittest.TestCase):
                     'artist': 1000
                     }
 
-        res = self.client().patch('/api/appointments/2', json=payload)
+        res = self.client().patch('/api/appointments/2', json=payload, headers={"Authorization": self.manager_jwt})
         data = json.loads(res.data)
 
         self.assertEqual(res.status_code, 404)
@@ -412,7 +415,7 @@ class TattooShopTestCase(unittest.TestCase):
                     'client': 1000
                     }
 
-        res = self.client().patch('/api/appointments/2', json=payload)
+        res = self.client().patch('/api/appointments/2', json=payload, headers={"Authorization": self.manager_jwt})
         data = json.loads(res.data)
 
         self.assertEqual(res.status_code, 404)
@@ -425,7 +428,7 @@ class TattooShopTestCase(unittest.TestCase):
     # Test deleting artist that does not exist return 404
     def test_delete_artist_error(self):
         
-        res = self.client().delete('/api/artists/1000')
+        res = self.client().delete('/api/artists/1000', headers={"Authorization": self.manager_jwt})
         data = json.loads(res.data)
 
         self.assertEqual(res.status_code, 404)
@@ -433,7 +436,7 @@ class TattooShopTestCase(unittest.TestCase):
         self.assertTrue(data['message'])
     # Test deleting client that does not exist return 404
     def test_delete_client_error(self):
-        res = self.client().delete('/api/clients/1000')
+        res = self.client().delete('/api/clients/1000', headers={"Authorization": self.manager_jwt})
         data = json.loads(res.data)
 
         self.assertEqual(res.status_code, 404)
@@ -442,13 +445,82 @@ class TattooShopTestCase(unittest.TestCase):
     
     # Test deleting appointment that does not exist return 404
     def test_delete_appointment_error(self):
-        res = self.client().delete('/api/appointments/1000')
+        res = self.client().delete('/api/appointments/1000', headers={"Authorization": self.manager_jwt})
         data = json.loads(res.data)
 
         self.assertEqual(res.status_code, 404)
         self.assertEqual(data['success'], False)
         self.assertTrue(data['message'])
     
+    '''
+    Test Role Based Authentication
+    '''
+
+    # Test Client can view a single appointment
+    def test_client_view_appointment(self):
+
+        res = self.client().get('/api/appointments/2',
+                                headers={
+                                         "Authorization": self.client_jwt
+                                         }
+                                )
+        data = json.loads(res.data)
+
+        self.assertEqual(res.status_code, 200)
+        self.assertEqual(data['success'], True)
+        self.assertEqual(data['appointment']['id'], 2)
+
+    # Test Client forbidden to update an appointment
+    def test_client_update_appointment_unauthorized(self):
+        
+        payload = {
+                    "artist": 2
+                   }
+        res = self.client().patch('/api/appointments/2',
+                                  json=payload,
+                                  headers={
+                                           "Authorization": self.client_jwt
+                                           }
+                                  )
+        data = json.loads(res.data)
+
+        self.assertTrue(res.status_code, 401)
+        self.assertEqual(data['success'], False)
+        self.assertTrue(data['message'])
+
+    # Test artist can update an appointment
+    def test_artist_update_appointment(self):
+        
+        payload = {
+                    'appointment_date': datetime(2021, 4, 6, 3, 30)
+                    }
+        res = self.client().patch('/api/appointments/2',
+                                  json=payload, 
+                                  headers={
+                                           "Authorization": self.artist_jwt
+                                           }
+                                  )
+        data = json.loads(res.data)
+        
+        self.assertEqual(res.status_code, 200)
+        self.assertEqual(data['success'], True)
+        self.assertEqual(datetime.strptime(data['appointment']['appointment_date'], "%a, %d %b %Y %I:%M:%S %Z"),
+                         payload['appointment_date']
+                         )
+
+    # Test artist unable to delete a client
+    def test_artist_delete_client_unauthorized(self):
+        
+        res = self.client().delete('/api/clients/2',
+                                   headers={
+                                            "Authorization": self.artist_jwt
+                                            }
+                                    )
+        data = json.loads(res.data)
+
+        self.assertEqual(res.status_code, 401)
+        self.assertEqual(data['success'], False)
+        self.assertTrue(data['message'])
 
 if __name__ == '__main__':
     unittest.main()
